Use a track's album-artist tag as the artist for the album.
If the album-artist isn't present, fallback to the track artist for the
album.  Also, the UI now searches based on the album and track artist id,
instead of the artist.

This means tracks whose artists who appear on albums for which they are not the
album-artist will now show up in the artists track listing.